improved send err msg. cleanup b4 login. new footer links

+++ src/services/device/cleanup.ts
@@ -1,10 +1,10 @@
 import { LocalDB } from '../../db';
 import { deleteDevice } from '../api';
 
-export default async function cleanupDevice() {
+export default async function cleanupDevice(callApiCleanup = true) {
 	try {
 		// continue even if remote device delete fails
-		await deleteDevice();
+		if (callApiCleanup) await deleteDevice();
 	} catch (e) {
 		console.error(e);
 	}
